refacotr(popular-films): refactor list-films into popular-films to avoid confusion
