+++ src/Controllers/login.js
@@ -6,7 +6,7 @@ const LoginUser = class LoginUser {
 
     const token = localStorage.getItem('tmdb_token');
     if (token) {
-      window.location.href = '/list-films';
+      window.location.href = '/popular-films';
       return;
     }
 
@@ -64,7 +64,7 @@ const LoginUser = class LoginUser {
 
       if (isValid) {
         localStorage.setItem('tmdb_token', token);
-        window.location.href = '/list-films';
+        window.location.href = '/popular-films';
       } else {
         input.classList.add('is-invalid');
         const feedback = document.querySelector('.invalid-feedback');
